Refactor some stuff to please codeclimate
License: MIT
